Updates after running FindBugs

+++ modules/application/src/main/java/com/sldeditor/tool/html/ExportHTML.java
@@ -172,7 +172,10 @@ public class ExportHTML
                 }
             }
 
-            file.delete();
+            if(file != null)
+            {
+                file.delete();
+            }
         }
     }
