+++ modules/application/src/test/java/com/sldeditor/test/unit/ui/detail/config/symboltype/externalgraphic/FieldConfigFilenameTest.java
@@ -298,8 +298,11 @@ public class FieldConfigFilenameTest {
 
         ExternalGraphicImpl externalGraphic = (ExternalGraphicImpl) styleBuilder.createExternalGraphic(filename, "png");
         field.setValue(fieldConfigManager, null, externalGraphic);
-        
-        f.delete();
+
+        if(f != null)
+        {
+            f.delete();
+        }
     }
 
     /**
